scsi: ufs-qcom: Remove pm_qos usage added by qcom
These pieces of codes affine a pm_qos before a transfer request is
called, then disables it after the UFS clock is turned off, leaving the
CPU in a shallow idle state for longer than necessary. Furthermore, this
affines the pm_qos to all cores, wasting immense amounts of power.
Remove the code entirely in order to replace it with a different
implementation.

Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>